+++ AO3Statistics/ExtensionMethods/HtmlNodeExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using HtmlAgilityPack;
 
-namespace AO3Statistics.ConsoleApp.ExtensionMethods;
+namespace AO3Statistics.ExtensionMethods;
 
 /// <summary>
 /// Extension methods for <see cref="HtmlNode"/>.
