+++ AO3Statistics/PasswordEncryptor.cs
@@ -1,7 +1,7 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
 
-namespace AO3Statistics.ConsoleApp;
+namespace AO3Statistics;
 
 internal static class PasswordEncryptor
 {
